refactor : BuyLotto 클래스명 변경 및 상수정의
BuyLotto 클래스명 LottoStore로 변경하여 돈을 입력받으면 구매한 로또 장수와 거스름돈을 계산
로또 가격 상수정의
